@@ -13,7 +13,7 @@
     "apexcharts": "^3.35.5",
     "clsx": "^2.0.0",
     "framer-motion": "^10.16.0",
-    "next": "13.4.16",
+    "next": "^14.0.4",
     "next-themes": "0.2.1",
     "react": "18.2.0",
     "react-apexcharts": "^1.4.0",
