@@ -1,14 +1,15 @@
+"use client";
 import { Button, Input } from "@nextui-org/react";
 import Link from "next/link";
 import React from "react";
-import { DotsIcon } from "../icons/accounts/dots-icon";
-import { ExportIcon } from "../icons/accounts/export-icon";
-import { InfoIcon } from "../icons/accounts/info-icon";
-import { TrashIcon } from "../icons/accounts/trash-icon";
-import { HouseIcon } from "../icons/breadcrumb/house-icon";
-import { UsersIcon } from "../icons/breadcrumb/users-icon";
-import { SettingsIcon } from "../icons/sidebar/settings-icon";
-import { TableWrapper } from "../table/table";
+import { DotsIcon } from "@/components/icons/accounts/dots-icon";
+import { ExportIcon } from "@/components/icons/accounts/export-icon";
+import { InfoIcon } from "@/components/icons/accounts/info-icon";
+import { TrashIcon } from "@/components/icons/accounts/trash-icon";
+import { HouseIcon } from "@/components/icons/breadcrumb/house-icon";
+import { UsersIcon } from "@/components/icons/breadcrumb/users-icon";
+import { SettingsIcon } from "@/components/icons/sidebar/settings-icon";
+import { TableWrapper } from "@/components/table/table";
 import { AddUser } from "./add-user";
 
 export const Accounts = () => {
